requirements.txt
streamlit>=1.30.0
chromadb>=0.4.22
openai>=1.10.0
anthropic>=0.18.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tiktoken>=0.5.0
